Adds helper class FileLoaderHelper
Description:
- Removes getInputStream method from AASXToMetamodelConverter
- Adds helper class FileLoaderHelper to facilitate loading of files
